tests/wpt/tests/content-security-policy/sandbox/shared-worker-sandbox.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
// Currently, sandbox directives for workers are not specified
// https://github.com/w3c/webappsec-csp/issues/279
// and thus this test asserts that the origin of SharedWorker is not sandboxed.
async_test(t => {
  const worker = new SharedWorker("support/sandboxed-shared-worker.js?" + Math.random());
  worker.onerror = t.unreached_func("SharedWorker construction failed");
  worker.port.onmessage = t.step_func_done(e => {
    assert_equals(e.data, location.origin, "Origin should not be sandboxed");
  });
}, "sandbox directive for SharedWorker");
</script>
</body>
